Fixed GRAMMARTESTS not semantictests. my bad.
