+++ src/test/grammartests.py
@@ -16,41 +16,13 @@ from semanticanalyzer import Semantic
 Token = lexer.Token
 LarkError = exceptions.LarkError
 
-class TestSignalGeneration(unittest.TestCase):
-   
-    def setUp(self):
-        self.l = lark.Lark(grammar.getgrammar(), parser='lalr', lexer="contextual")
-        self.measureTest = '''
-        Compose {
-            Measure {
-                acousticgrandpiano {
-                    1/2 C4; 1/2 --;
-                }
-                trumpet {
-                    1/4 E4; 1/4 (C4 E4); 1/4 (E4 G4); fff; 1/4 tuplet(C4 D4 E4 F4 G4); 
-                }
-            }
-        }         
-        '''
-
-        self.tree = self.l.parse(self.measureTest)
-        self.testMeasure = self.tree.children[0].children[0].children[0]
-
-        # what args go here???
-        self.semantic = Semantic(self.testMeasure)
-        self.help = TestHelp()
-
-    def test_measure_to_signal(self):
-        self.semantic.measure_to_signal(self.testMeasure)
-        
-
 class TestGrammar(unittest.TestCase):
 
     def setUp(self):
         self.l = lark.Lark(grammar.getgrammar(), parser='lalr', lexer="contextual")
         self.help = TestHelp()
 
-    
+
     def test_variable(self):
         test = '''
         compose{
@@ -114,7 +86,7 @@ class TestGrammar(unittest.TestCase):
                         accidental b
                         number 4
         '''
-        
+
         testtree = self.l.parse(test).pretty()
         print(self.l.parse(test))
         self.assertTrue(self.help.prettyTreeComp(testtree, accept), 'Basic Measure syntax parsed incorrectly')
@@ -138,15 +110,15 @@ class TestGrammar(unittest.TestCase):
         '''
         testtree = self.l.parse(test).pretty()
         self.assertTrue(self.help.prettyTreeComp(testtree, accept), 'Basic repeat tree incorrect')
-        
-        
+
+
     def test_repeatFail(self):
         test='''
         Compose {
             Repeat
             Tempo(72)
         }
-        ''' 
+        '''
         reject = '''
         start
             compose
